+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 + [x] Windows: fix interop between Metro4 and jQuery
 + [x] VideoPlayer: fix interop between Metro4 and jQuery
 + [x] AudioPlayer: fix interop between Metro4 and jQuery
++ [x] Hotkey: fix interop between Metro4 and jQuery
 + [x] TimePicker: fix interop between Metro4 and jQuery
 + [x] SideBar: fix interop between Metro4 and jQuery
 + [x] Charms: fix interop between Metro4 and jQuery